+++ localization/enUS.lua
@@ -2,6 +2,7 @@ local L = {}
 local _G = _G
 
 _G["BINDING_NAME_MOUNTY_MAGIC"] = "Summon magic mount"
+_G["BINDING_NAME_MOUNTY_GROUND"] = "Summon ground mount"
 _G["BINDING_NAME_MOUNTY_WATER"] = "Summon water mount"
 _G["BINDING_NAME_MOUNTY_TAXI"] = "Summon taxi mount"
 _G["BINDING_NAME_MOUNTY_REPAIR"] = "Summon repair mount"
@@ -21,9 +22,10 @@ L["mode.Random"] = "Random"
 L["Taxi!"] = "Taxi!"
 
 L["Options"] = "Mounty #TLV - Options"
-L["Random"] = "Random"
-L["Don't show off in resting areas"] = "Don't automatically show off in resting areas"
-L["Don't fly (except if taxi)"] = "Don't fly in group (except as a taxi)"
+L["Quick start"] = "Quick start"
+L["Random"] = "Random (If not checked, each category will cycle through)"
+L["Look at me!"] = "Look at me! Show off in resting areas"
+L["Stay together"] = "Stay together and don't fly in group (except as a taxi)"
 L["Taxi mode"] = "Taxi mode"
 L["Debug mode"] = "Debug mode"
 L["Auto open"] = "Auto open and close with mount journal"
@@ -33,14 +35,18 @@ L["How to call a passenger"] = "How to call a passenger"
 L["Helptext"] = "Drag mounts into here. Right click to remove a mount."
 
 L["Debug: "] = "Debug: "
-L["fly mode: "] = "fly mode: "
-L["random: "] = "random: "
-L["taxi: "] = "taxi: "
+L["Auto open & close: "] = "Auto open & close: "
+L["Together mode: "] = "Together mode: "
+L["Show off mode: "] = "Show off mode: "
+L["Random: "] = "Random: "
+L["Taxi mode: "] = "Taxi mode: "
 
 L["on"] = "on"
 L["off"] = "off"
 
 L["Open Mounty"] = "Mounty #TLV"
+
+L["Quick start full"] = "- Pull your favourite mounts from the mount journal into the categories of Mounty #TLV.\n- Open WoW's settings for key bindings, select Mounty #TLV and configure your magic key to mount.\n- Press your magic key to pick up the perfect mount for here and now.\n- Enjoy Mounty #TLV and take a look and all the other options and possibilities."
 
 local _, Mounty = ...
 Mounty.L = L